Enable more compiler warnings
We *are* security-oriented project so let's enable as many relevant
flags as possible. Here are some flags that I found useful after
looking through the latest lists for GCC 8 and Clang 7. (In addition
to catch-all "-Wall -Wextra" which enable most of the warnings.)

Not all flags are supported by all modern compiler versions that are
still in use. Wrap all relatively new flags into 'supported' checks
which verify that the flag seems to be actually supported by the
compiler. Skip all flags thar are not supported: we're not going to
get any use from them anyway but this should not prevent compilation.