Strict function prototypes in C code
Recently added "-Wstrict-prototypes" has found a bunch of functions
(in unit tests) which do not have strict prototypes. In C functions
that have "()" argument lists and do not have an explicit prototype
are allowed to accept arbitrary amount of arguments.

Ensure that all nullary functions have a "(void)" argument list for
consistency with many other places which already use this spelling.