Prevent a possible buffer overflow in ThemisPP
This is an example of a bug which is prevented by "-Wunused-parameters"
and "-Wunused-variables" warnings. Here these warnings point out that
the "data_length" parameter is never used anywhere in the function
while in fact it should be used in buffer length checks. If the C++
implementation returns a vector longer than the allocated buffer then
the following memcpy() call will cause a buffer overflow.