Use special compiler flags for AFL-enabled builds
It turns out that afl-clang tends to have a completely different set
of supported flags than normal clang and gcc. Therefore we have to be
more picky about the flags we set in AFL_CFLAGS. Filter out all flags
that AFL_CC does not seem to support.

Teach "supported" function to accept an optional second argument
that sets the compiler to test. It will use $(CC) by default if
the argument is omitted.